refactors the code base to use a new system of test execution, updates a variety of automation scripts
