+++ regression-tests/automation/cleanup.sh
