+++ regression-tests/automation/install_qa.sh
@@ -4,5 +4,3 @@ NAME="[`basename $0`]"
 DIR="$( cd "$( dirname "$0" )" && pwd )"
 echo "$NAME DIR=$DIR"
 
-
-
