+++ regression-tests/automation/install_source.sh
