+++ regression-tests/sparktkregtests/testcases/frames/confusion_matrix_test.py
@@ -1,11 +1,7 @@
 """Tests the confusion matrix functionality against known values"""
 import unittest
 
-import os
-import sys
-sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.realpath(__file__)))))
-
-from qalib import sparktk_test
+from sparktkregtests.lib import sparktk_test
 
 
 class ConfusionMatrix(sparktk_test.SparkTKTestCase):
