--- regression-tests/utils/home.bash
@@ -1,10 +1,0 @@
-#!/bin/bash
-NAME="[`basename ${BASH_SOURCE[0]}`]"
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-
-# echo "$NAME DIR=$DIR"
-
-MAINDIR="$(dirname $DIR)"
-MAINDIR="$(dirname $MAINDIR)"
-
-export SPARKTK_HOME=$MAINDIR/core/target
